Fix underflow

+++ contracts/strategies/operations/Operation.sol
@@ -128,6 +128,6 @@ abstract contract Operation is IOperation {
     function _normalizeDecimals(address _asset, uint256 _quantity) internal view returns (uint256) {
         uint8 tokenDecimals = ERC20(_asset).decimals();
         require(tokenDecimals <= 18, 'Unsupported decimals');
-        return tokenDecimals != 18 ? _quantity.mul(10**(18.sub(tokenDecimals))) : _quantity;
+        return tokenDecimals != 18 ? _quantity.mul(10**(uint256(18).sub(tokenDecimals))) : _quantity;
     }
 }
